Add working search City
