+++ src/data/Cities.js
@@ -1,0 +1,60 @@
+import towerImage from "../../assets/images/Tower.jpg";
+import saintImage from "../../assets/images/Saint.jpg";
+import adduImage from "../../assets/images/Addu.jpg";
+import marinaImage from "../../assets/images/Marina.jpg";
+import charlesImage from "../../assets/images/Charles.jpg";
+
+export const CITIES = [
+    {
+      id: "0",
+      city: "Saint Moritz",
+      image: saintImage,
+      price: "$1200",
+      persons: "person",
+      location: "Zurich, Switzerland",
+      rating: "5.0",
+      review: "6",
+      author: "Leila Azevedo",
+    },
+    {
+      id: "1",
+      city: "Tower Bridge",
+      image: towerImage,
+      price: "$890",
+      persons: "person",
+      location: "London, United Kingdom",
+      rating: "5.0",
+      review: "3.5",
+      author: "David Monaghan",
+    },
+    {
+      id: "2",
+      city: "Addu Atoll",
+      image: adduImage,
+      price: "$900",
+      persons: "person",
+      location: "Maldives",
+      rating: "5.0",
+      review: "1.7",
+    },
+    {
+      id: "3",
+      city: "Marina Bay",
+      image: marinaImage,
+      price: "$750",
+      persons: "person",
+      location: "Singapore",
+      rating: "5.0",
+      review: "2.5",
+    },
+    {
+      id: "4",
+      city: "Charles Bridge",
+      image: charlesImage,
+      price: "$700",
+      persons: "person",
+      location: "Prague",
+      rating: "5.0",
+      review: "2.0",
+    },
+  ];
