+++ src/data/FilterCities.js
@@ -1,0 +1,5 @@
+import {CITIES} from "../data/Cities.js"
+
+export const filterCitiesByReview = (isGreaterThan) => {
+  return CITIES.filter((city) => isGreaterThan ? Number(city.review) > 3 : Number(city.review) <= 3);
+};
